+++ src/main/java/io/github/astrapi69/crypto/key/writer/KeyWriter.java
@@ -32,7 +32,7 @@ import java.util.Objects;
 
 import org.bouncycastle.openssl.jcajce.JcaPEMWriter;
 
-import de.alpharogroup.file.write.WriteFileExtensions;
+import io.github.astrapi69.write.WriteFileExtensions;
 
 /**
  * The class {@link KeyWriter} is a utility class for write security keys in files.