+++ src/test/java/io/github/astrapi69/crypto/blockchain/TransactionTest.java
@@ -37,8 +37,8 @@ import org.meanbean.test.BeanTester;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-import de.alpharogroup.evaluate.object.evaluators.EqualsHashCodeAndToStringEvaluator;
-import de.alpharogroup.file.search.PathFinder;
+import io.github.astrapi69.evaluate.object.evaluators.EqualsHashCodeAndToStringEvaluator;
+import io.github.astrapi69.search.PathFinder;
 import io.github.astrapi69.crypto.key.reader.PublicKeyReader;
 
 /**