+++ src/test/java/io/github/astrapi69/crypto/key/PrivateKeyExtensionsTest.java
@@ -46,7 +46,7 @@ import org.meanbean.test.BeanTester;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-import de.alpharogroup.file.search.PathFinder;
+import io.github.astrapi69.search.PathFinder;
 import io.github.astrapi69.crypto.algorithm.KeyPairGeneratorAlgorithm;
 import io.github.astrapi69.crypto.factories.KeyPairFactory;
 import io.github.astrapi69.crypto.key.reader.PrivateKeyReader;