+++ src/test/java/io/github/astrapi69/crypto/key/reader/PublicKeyReaderTest.java
@@ -34,7 +34,7 @@ import org.bouncycastle.jce.provider.BouncyCastleProvider;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
-import de.alpharogroup.file.search.PathFinder;
+import io.github.astrapi69.search.PathFinder;
 import io.github.astrapi69.crypto.key.PublicKeyExtensions;
 
 /**